Write progress to stderr
