+++ src/docfx/docset/Docset.cs
@@ -79,7 +79,7 @@ namespace Microsoft.Docs.Build
 
         private HashSet<Document> CreateBuildScope(IEnumerable<Document> redirections)
         {
-            using (Log.Measure("Globbing files"))
+            using (Progress.Start("Globbing files"))
             {
                 return FileGlob.GetFiles(DocsetPath, Config.Content.Include, Config.Content.Exclude)
                                .Select(file => Document.TryCreateFromFile(this, Path.GetRelativePath(DocsetPath, file)))
